Reportモデル、ReportViewモデル作成

+++ src/main/java/actions/EmployeeAction.java
@@ -25,6 +25,7 @@ public class EmployeeAction extends ActionBase {
 
 
     public void index() throws ServletException, IOException{
+        if (checkAdmin()) {
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
@@ -39,20 +40,22 @@ public class EmployeeAction extends ActionBase {
         if(flush != null) {
             putRequestScope(AttributeConst.FLUSH, flush);
             removeSessionScope(AttributeConst.FLUSH);
-        }
         forward(ForwardConst.FW_EMP_INDEX);
+        }
+        }
     }
 
     public void entryNew() throws ServletException, IOException{
+        if (checkAdmin()) {
         putRequestScope(AttributeConst.TOKEN, getTokenId());
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
-
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     public void create() throws ServletException, IOException {
 
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             EmployeeView ev = new EmployeeView(
                     null,
@@ -87,6 +90,7 @@ public class EmployeeAction extends ActionBase {
     }
 
     public void show() throws ServletException, IOException{
+        if (checkAdmin()) {
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
         if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
@@ -97,9 +101,11 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     public void edit() throws ServletException, IOException{
+        if (checkAdmin()) {
 
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
         if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()){
@@ -110,11 +116,12 @@ public class EmployeeAction extends ActionBase {
         putRequestScope(AttributeConst.EMPLOYEE, ev);
 
         forward(ForwardConst.FW_EMP_EDIT);
+        }
     }
 
     public void update() throws ServletException, IOException{
 
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) {
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
                     getRequestParam(AttributeConst.EMP_CODE),
@@ -142,10 +149,20 @@ public class EmployeeAction extends ActionBase {
     }
 
     public void destroy() throws ServletException, IOException{
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) {
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
             putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    private boolean checkAdmin() throws ServletException, IOException {
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }else {
+            return true;
         }
     }
 }
